改善 SQL Server 相容性

+++ database/migrations/2017_12_11_053958_create_ip2nation_table.php
@@ -14,7 +14,7 @@ class CreateIp2nationTable extends Migration
     public function up()
     {
         Schema::create('ip2nation', function (Blueprint $table) {
-            $table->integer('ip')->default(0);
+            $table->bigInteger('ip')->default(0);
             $table->char('country', 2);
         });
     }
